fix: Move runtime codes to own separate header file (fixes #697)
